requirements-common.txt
alembic==0.7.3
celery==3.1.17
jac==0.14.1
psycopg2==2.5.2
python-firebase==1.2
python-memcached==1.53
pytz==2014.9
rauth==0.7.0
requests==2.5.3
rjsmin==1.0.10
SQLAlchemy==0.9.8
twython==3.2.0
WTForms==2.0.1
WTForms-JSON==0.2.7

Flask==0.10.1
Flask-Login==0.2.11
Flask-SeaSurf==0.1.22
Flask-SQLAlchemy==2.0
